% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoter-package.R
\docType{package}
\name{zoter-package}
\alias{zoter}
\alias{zoter-package}
\title{zoter: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kleinlennart/zoter}
  \item \url{https://kleinlennart.github.io/zoter/}
  \item Report bugs at \url{https://github.com/kleinlennart/zoter/issues}
}

}
\keyword{internal}
