% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_zotero.R
\name{req_zotero}
\alias{req_zotero}
\title{Title}
\usage{
req_zotero(
  query = NULL,
  output = c("parsed", "raw"),
  sort = "dateAdded",
  format = c("json", "keys", "versions"),
  limit = 100,
  paginate = TRUE,
  verbose = FALSE,
  interactive = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
